refactor - move boradgame and player softdelete actions inside a concern
Not 100% generic concern, and coupled to models, but helps to clarify models purpose.
Consider to clean really_destroy_when_useless in a future and free_to_loan? less dependent of external actions in a future.

+++ app/models/player.rb
@@ -1,5 +1,6 @@
 class Player < ApplicationRecord
-  acts_as_paranoid
+  include SoftDeletable
+
   has_and_belongs_to_many :events
   has_and_belongs_to_many :organizations, -> { distinct }
   has_many :loans
@@ -14,45 +15,31 @@ class Player < ApplicationRecord
                     format: { with: VALID_EMAIL_REGEX },
                     uniqueness: { case_sensitive: false }
 
-  before_destroy :not_removed_with_pending_loans, :mark_fields_as_deleted
-  after_destroy :really_destroy_when_useless, :remove_future_participants
+  after_destroy :remove_future_participants
 
   def name
-    firstname + " " + lastname
+    "#{firstname} #{lastname}"
   end
 
   def dni_plus_name
-    dni + " | " + name
+    "#{dni} | #{name}"
   end
 
   def age
-    return "" unless birthday
+    return '' unless birthday
     today = Date.today
     had_birthday = today.month > birthday.month || (today.month == birthday.month && today.day >= birthday.day) ? true : false
     today.year - birthday.year - (had_birthday ? 0 : 1)
   end
 
   def active_loans
-    loans.where(returned_at: nil).count
+    loans.where(returned_at: nil)
   end
 
   private
 
-  def mark_fields_as_deleted
-    update_columns(firstname: "DELETED", lastname: "DELETED", email: "DELETED", dni: "DELETED")
-  end
-
-  def not_removed_with_pending_loans
-    errors.add(:destroy, "Cannot delete players with pending loans") unless active_loans.zero?
-    throw(:abort) unless active_loans.zero?
-  end
-
-  def really_destroy_when_useless
-    really_destroy! unless loans.present? || participants.present? || frozen?
-  end
-
   def remove_future_participants
-    false unless participants.present?
+    return false unless participants.present?
     participants.each do |participant|
       participant.destroy if participant.at_future_tournament?
     end
